Restructure import sequence of Sykle __version__ in setup.py to avoid importing 3rd party dependencies during installation.
Noticed that attempting to install Sykle with pip results in an error
due to a missing `dotenv` dependency. Thus far, I believe we've been
chalking this up to some sort of fluke with how `install_requires` is
defined in setup.py, and working around the issue by manually
pre-installing python-dotenv (and other dependencies) in environments
where we install Sykle. Took a closer look and noticed that setup.py is
importing the `__version__` from __init__.py. This would be fine, except
that in __init__.py we're importing the `Sykle` class from `sykle.py`,
which in turn calls for the import of dotenv, among other modules. This
commit flips the structure so that the version is define in __init__.py
and pulled into the Sykle class, rather than the reverse.